Screen2
organised styles into separate files, imported lock image onto screen2
