+++ App.js
@@ -1,20 +1,9 @@
-import { StatusBar } from 'expo-status-bar';
-import { StyleSheet, Text, View } from 'react-native';
+import Container from './styles/Container';
 
 export default function App() {
   return (
-    <View style={styles.container}>
-      <Text>Testing</Text>
-      <StatusBar style="auto" />
-    </View>
+    <Container>
+    </Container>
   );
 }
 
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#fff',
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-});
