+++ package.json
@@ -14,7 +14,9 @@
     "react": "18.2.0",
     "react-native": "0.72.4",
     "react-native-web": "~0.19.6",
-    "react-dom": "18.2.0"
+    "react-dom": "18.2.0",
+    "typescript": "^5.1.3",
+    "@types/react": "~18.2.14"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0"
