+++ tsconfig.json
@@ -1,0 +1,5 @@
+{
+  "extends": "@tsconfig/react-native/tsconfig.json",
+  "compilerOptions": {
+    "jsx": "react"}
+}
